scoring-calculator/requirements.txt
uvicorn==0.14.0
fastapi==0.65.2
redis==3.5.3
pymongo==3.11.4
mongoengine==0.23.1
pydantic==1.8.2
dateutils==0.6.12
jdatetime==3.6.2
kafka-python==2.0.2
pytest